msmfb/ts/vkeyb/fb_helper: ts now uses interrupts instead of a polling mechanism
The touchscreen driver has been updated to use interrupts instead of a polling mechanism. However, the 'pen up' event does not generate a reliable interrupt. Due to this behaviour there is still a timeout
mechanism that sends the 'pen up' event manually when no interrupts were fired within 200ms of the last 'pen down' interrupt.

Removing the polling mechanism from the touchscreen driver had some impact as the virtual keyboard depended on it.

The following (non-trivial) changes have been made to fix this:
1) The fb_helper.h has been added to centralize painting on the framebuffer (ts uses it for calibration, vkeyb for.. the vkeyb). The fb_helper contains generic code and is not msm_fb specific.
2) A callback mechanism has been added to msm_fb to allow drivers to do last-millisecond updates to the framebuffer right before the DMA takes place. Check msm_fb.h on how to use this.