msmfb: added refresh thread to support X's fbdev (disabled by default)
As the touchscreen is now interrupt based instead of polling, the screen does not get updated in a regular fashion. The refresh thread has been added to compensate for this. As it's not a very nice solution,
it is disabled by default and can be enabled at compile-time by setting CONFIG_MSM_FB_REFRESH.

Note that this refresh thread is not neccesary for Android.