mem: fixed the memory map and appy fixup for memory banks
The memory map, defined in the board-htcraphael.h file, has been fixed. Parts of memory were unusable as they seem to be locked by the MPU (1MB at start of SMI and last 13MB of first sdram bank).

The fixup is used to prevent being dependent on what HaRET passes as amount of memory to be more on the safe side.

Note that the second memory bank is currently disabled for two reasons:
1) not all devices have this memory bank, so there needs to be some sort of detection mechanism
2) it somehow hangs the devices when trying to start android (hangs even before the animated splash screen is shown)

The VMALLOC_END has been adjusted as it would overlap 128MB of the second memory bank (which is all of it). This is in preparation of enabling the second memory bank.

The Android ram console has been disabled. It's not used and saves us another 128KB ;-)